requirements.txt
nicegui
pyserial